OSIS-927 Move proposal/common
