+++ base/forms/learning_unit_proposal.py
@@ -29,8 +29,8 @@ from itertools import chain
 from django import forms
 from django.db import transaction
 
-from base.business.learning_unit_proposal import reinitialize_data_before_proposal, compute_proposal_type
-from base.business.learning_units.proposal.common import compute_proposal_state
+from base.business.learning_unit_proposal import reinitialize_data_before_proposal, compute_proposal_type, \
+    compute_proposal_state
 from base.forms.learning_unit.learning_unit_create import EntitiesVersionChoiceField
 from base.forms.learning_unit.learning_unit_create_2 import FullForm, PartimForm
 from base.models import entity_container_year
